fix sand mention

+++ app/Jobs/SandMention.php
@@ -33,7 +33,7 @@ class SandMention implements ShouldQueue
      */
     public function __construct(
         readonly Car     $car,
-        readonly ?string $path,
+        readonly string  $path,
         readonly ?string $description,
         readonly User    $user,
         readonly Carbon  $time
@@ -46,13 +46,15 @@ class SandMention implements ShouldQueue
      */
     public function handle(): void
     {
-        if ($this->path) {
+        $isFile = $this->path !== '';
+        if ($isFile) {
             $storagePath = storage_path('app/private/' . $this->path);
             $file = new UploadedFile($storagePath, basename($storagePath), mime_content_type($storagePath), null, true);
         }
         $mention = MentionEloquent::create($this->car, $this->user, $this->description, $file ?? null);
 
-        Storage::delete($this->path);
+        if ($isFile)
+            Storage::delete($this->path);
 
         $imageUrl = $mention->getFirstMedia(EnumTypeMedia::PHOTO_MENTION->value)?->getPath();
         $text = TelegramBotHelpers::generationTextMention($this->user, $this->car, $this->description, $this->time);
